implementada Classe de cadastro de fornecedores com validacoes e novos campos, adicionadas tabela e campos na classe de criacao do banco de dados

+++ OPS_OphellSystem/Cadastros/Classes/PessoaJuridica.cs
@@ -5,7 +5,7 @@
         #region "Variaveis"
         protected string _fantasia;
         protected string _razao;
-        protected int _cnpj;
+        protected long _cnpj;
         protected string _endereco;
         protected int _numero;
         protected string _cidade;
